+++ EasyDriver/EasyDriver/DriverApi.cs
@@ -4,18 +4,38 @@ using OpenQA.Selenium;
 
 namespace Comfast.EasyDriver;
 
+/// <summary>
+/// Main API of framework
+/// </summary>
 public static class DriverApi {
-    public static IWebDriver Driver => Configuration.GetDriver();
-    public static string CurrentUrl => Driver.Url;
+    /// <summary>
+    /// Returns current WebDriver instance.
+    /// </summary>
+    public static IWebDriver GetDriver() => Configuration.DriverProvider.GetDriver();
 
-    public static SimpleLocator S(string cssOrXpath) => new(cssOrXpath);
-    public static void NavigateTo(string url) => Driver.Navigate().GoToUrl(url);
+    /// <summary>
+    /// Shortcut to navigation
+    /// </summary>
+    public static void NavigateTo(string url) => GetDriver().Navigate().GoToUrl(url);
 
+    /// <summary>
+    ///
+    /// </summary>
+    /// <param name="cssOrXpath">Any selector</param>
+    /// <param name="description"></param>
+    public static SimpleLocator S(string cssOrXpath, string? description = null) => new(cssOrXpath, description);
+
+    /// <summary>
+    /// Executes JavaScript from browser console.
+    /// </summary>
     public static T ExecuteJs<T>(string jsCode, params object[] args) {
-        var jsDriver = (IJavaScriptExecutor)Driver;
+        var jsDriver = (IJavaScriptExecutor)GetDriver();
         return (T)jsDriver.ExecuteScript(jsCode, args);
     }
 
+    /// <summary>
+    /// Wait for any action to return true. Ignore Exceptions.
+    /// </summary>
     public static void WaitFor(Func<bool> action, string? description = null, int? timeoutMs = null) {
         WaitUtils.WaitFor(action, description, timeoutMs);
     }