+++ EasyDriver/EasyDriver/Se/Infra/Browser/SimpleBrowserRunner.cs
@@ -1,0 +1,12 @@
+﻿using Comfast.EasyDriver.Models;
+using OpenQA.Selenium;
+
+namespace Comfast.EasyDriver.Se.Infra.Browser;
+
+internal class SimpleBrowserRunner : IBrowserRunner {
+    private readonly Func<IWebDriver> _runFunc;
+    public SimpleBrowserRunner(Func<IWebDriver> runFunc) {
+        _runFunc = runFunc;
+    }
+    public IWebDriver RunNewBrowser() => _runFunc();
+}