+++ EasyDriver/EasyDriver/Se/Infra/DriverProvider.cs
@@ -1,8 +1,6 @@
 ﻿using Comfast.EasyDriver.Models;
+using Comfast.EasyDriver.Se.Infra.Browser;
 using OpenQA.Selenium;
-using OpenQA.Selenium.Chrome;
-using OpenQA.Selenium.Edge;
-using OpenQA.Selenium.Firefox;
 
 namespace Comfast.EasyDriver.Se.Infra;
 
@@ -12,12 +10,16 @@ namespace Comfast.EasyDriver.Se.Infra;
 /// - autoClose - close browser after process exit
 /// </summary>
 public class DriverProvider : IDriverProvider {
-    private readonly ThreadLocal<WebDriver> _instances;
+    private readonly ThreadLocal<IWebDriver> _instances;
     private readonly DriverConfig _driverConfig;
+    private readonly DriverSessionStore _driverSessionStore = new();
+
+    public IBrowserRunner BrowserRunner { get; set; }
 
     public DriverProvider(DriverConfig driverConfig) {
-        _instances = new ThreadLocal<WebDriver>(ProvideDriverInstance, true);
+        _instances = new ThreadLocal<IWebDriver>(ProvideDriverInstance, true);
         _driverConfig = driverConfig;
+        BrowserRunner = new BrowserRunner(driverConfig);
     }
 
     /// <summary>
@@ -40,17 +42,17 @@ Reconnect feature isn't compatible with parallel runs. Possible solutions:
     /// <summary>
     /// Run/reconnect to Browser instance
     /// </summary>
-    private WebDriver ProvideDriverInstance() {
+    private IWebDriver ProvideDriverInstance() {
         bool reconnect = _driverConfig.Reconnect;
         bool autoClose = _driverConfig.AutoClose;
 
-        var driver = reconnect && !autoClose
-            ? DriverSessionStore.RestoreSessionOrElse(RunNewDriver)
-            : RunNewDriver();
+        var webDriver = reconnect && !autoClose
+            ? _driverSessionStore.RestoreSessionOrElse(BrowserRunner.RunNewBrowser)
+            : BrowserRunner.RunNewBrowser();
 
-        if (autoClose) AddShutdownHook(driver);
+        if (autoClose) AddShutdownHook(webDriver);
 
-        return driver;
+        return webDriver;
     }
 
     /// <summary>
@@ -61,53 +63,5 @@ Reconnect feature isn't compatible with parallel runs. Possible solutions:
             driver.Close(); // closes browser
             driver.Dispose(); // closes driver process ( e.g. chromedriver.exe )
         };
-    }
-
-    /// <summary>
-    ///  Run new WebDriver instance
-    /// </summary>
-    private WebDriver RunNewDriver() {
-        var browser = _driverConfig.BrowserName;
-        switch (browser) {
-            case "chrome":
-            case "chromium":
-            case "brave":
-                return RunChromium();
-            case "firefox":
-                return RunFirefox();
-            case "edge":
-                return RunEdge();
-            default: throw new Exception("Invalid browser: " + browser);
-        }
-
-    }
-
-    private ChromeDriver RunChromium() {
-        var opts = new ChromeOptions();
-        if (_driverConfig.Headless) opts.AddArguments("headless");
-        opts.BinaryLocation = _driverConfig.BrowserPath;
-
-        try {
-            return new ChromeDriver(_driverConfig.DriverPath, opts);
-        } catch (InvalidOperationException e) {
-            if (e.Message.Contains("No process is associated with this object."))
-                throw new Exception("Invalid driver path: " + Configuration.DriverConfig.DriverPath, e);
-            throw;
-        }
-    }
-
-    private FirefoxDriver RunFirefox() {
-        var opts = new FirefoxOptions();
-        if (_driverConfig.Headless) opts.AddArguments("--headless");
-        opts.BinaryLocation = _driverConfig.BrowserPath;
-
-        return new FirefoxDriver(opts);
-    }
-
-    private EdgeDriver RunEdge() {
-        var opts = new EdgeOptions();
-        if (_driverConfig.Headless) opts.AddArguments("headless");
-        opts.BinaryLocation = _driverConfig.BrowserPath;
-        return new EdgeDriver(_driverConfig.DriverPath, opts);
     }
 }